Refactor msgraph-cli-core
