+++ src/msgraph-cli-core/msgraph/__init__.py
@@ -1,0 +1,2 @@
+import pkg_resources
+pkg_resources.declare_namespace(__name__)
